refactor: better class name
