+++ src/helpers/scope-validator.js
